v1.0.1
Updated example
